Add 180s timeout for captive portal and reboot if timeout is hit
During power outage the previously saved WiFi network is not
immediately available and thus ESPs have failed to connect and
have ended up stuck to captive portal waiting for configuration.

With the timeout if the configuration is not done via portal in 180s,
ESP is restarted and it tries to connect to previously saved network again.